backend/requirements.txt
flask
flask-cors
mysql-connector-python
pyjwt
bcrypt
python-dotenv

